requirements.txt
asgiref==3.11.1
Django==6.0.4
django-environ==0.13.0
gunicorn==25.3.0
packaging==26.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2026.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pytest==8.1.1
pytest-django==4.8.0
drf-spectacular==0.27.2
whitenoise==6.6.0
